Follow this configgery style for ease.
- or i'll just keep pull requesting it you heathen.
