+++ addons/opf/CfgFactionClasses.hpp
@@ -1,0 +1,8 @@
+class cfgFactionClasses {
+    class tg_opf_faction {
+        displayName = "Trench's Units";
+        priority = 3; // Position in list.
+        side = 1; // Opfor = 0, Blufor = 1, Indep = 2.
+        icon = ""; //Custom Icon
+    };
+};
