+++ addons/opf/CfgWeapons.hpp
@@ -1,0 +1,244 @@
+class cfgWeapons {
+    class ItemCore;
+    class UniformItem;
+    class Uniform_Base;
+
+    // Uniforms
+
+    // Polo & Khakis
+    class tg_u_gorka_afghan: Uniform_Base {
+        scope = 2;
+        displayName = "Gorka (Afghan)";
+		picture = "\tg\opf\Data\logo_small.paa";
+        model = "\A3\Characters_F\Common\Suitpacks\suitpack_civilian_F.p3d";
+        class ItemInfo : UniformItem {
+			uniformModel = "-";
+            uniformClass = "tg_opf_gorka_afghan"; // name of linked character
+            containerClass = "Supply50";
+            mass = 50;
+        };
+    };
+
+	class tg_u_gorka_blue_black: Uniform_Base {
+        scope = 2;
+        displayName = "Gorka (Blue/Black)";
+		picture = "\tg\opf\Data\logo_small.paa";
+        model = "\A3\Characters_F\Common\Suitpacks\suitpack_civilian_F.p3d";
+        class ItemInfo : UniformItem {
+			uniformModel = "-";
+            uniformClass = "tg_opf_gorka_blue_black"; // name of linked character
+            containerClass = "Supply50";
+            mass = 50;
+        };
+    };
+
+	class tg_u_gorka_blue: Uniform_Base {
+        scope = 2;
+        displayName = "Gorka (Blue)";
+		picture = "\tg\opf\Data\logo_small.paa";
+        model = "\A3\Characters_F\Common\Suitpacks\suitpack_civilian_F.p3d";
+        class ItemInfo : UniformItem {
+			uniformModel = "-";
+            uniformClass = "tg_opf_gorka_blue"; // name of linked character
+            containerClass = "Supply50";
+            mass = 50;
+        };
+    };
+
+	class tg_u_gorka_cbrn_black: Uniform_Base {
+        scope = 2;
+        displayName = "Gorka (CBRN Black)";
+		picture = "\tg\opf\Data\logo_small.paa";
+        model = "\A3\Characters_F\Common\Suitpacks\suitpack_civilian_F.p3d";
+        class ItemInfo : UniformItem {
+			uniformModel = "-";
+            uniformClass = "tg_opf_gorka_cbrn_black"; // name of linked character
+            containerClass = "Supply50";
+            mass = 50;
+        };
+    };
+
+	class tg_u_gorka_cbrn_black_alt: Uniform_Base {
+        scope = 2;
+        displayName = "Gorka (CBRN Black) (alt)";
+		picture = "\tg\opf\Data\logo_small.paa";
+        model = "\A3\Characters_F\Common\Suitpacks\suitpack_civilian_F.p3d";
+        class ItemInfo : UniformItem {
+			uniformModel = "-";
+            uniformClass = "tg_opf_gorka_cbrn_black_alt"; // name of linked character
+            containerClass = "Supply50";
+            mass = 50;
+        };
+    };
+
+	class tg_u_gorka_cbrn: Uniform_Base {
+        scope = 2;
+        displayName = "Gorka (CBRN)";
+		picture = "\tg\opf\Data\logo_small.paa";
+        model = "\A3\Characters_F\Common\Suitpacks\suitpack_civilian_F.p3d";
+        class ItemInfo : UniformItem {
+			uniformModel = "-";
+            uniformClass = "tg_opf_gorka_cbrn"; // name of linked character
+            containerClass = "Supply50";
+            mass = 50;
+        };
+    };
+
+	class tg_u_gorka_grey: Uniform_Base {
+        scope = 2;
+        displayName = "Gorka (Grey)";
+		picture = "\tg\opf\Data\logo_small.paa";
+        model = "\A3\Characters_F\Common\Suitpacks\suitpack_civilian_F.p3d";
+        class ItemInfo : UniformItem {
+			uniformModel = "-";
+            uniformClass = "tg_opf_gorka_grey"; // name of linked character
+            containerClass = "Supply50";
+            mass = 50;
+        };
+    };
+
+	class tg_u_gorka_klmk: Uniform_Base {
+        scope = 2;
+        displayName = "Gorka (KLMK)";
+		picture = "\tg\opf\Data\logo_small.paa";
+        model = "\A3\Characters_F\Common\Suitpacks\suitpack_civilian_F.p3d";
+        class ItemInfo : UniformItem {
+			uniformModel = "-";
+            uniformClass = "tg_opf_gorka_klmk"; // name of linked character
+            containerClass = "Supply50";
+            mass = 50;
+        };
+    };
+
+	class tg_u_gorka_klmk_alt: Uniform_Base {
+        scope = 2;
+        displayName = "Gorka (KLMK) (alt)";
+		picture = "\tg\opf\Data\logo_small.paa";
+        model = "\A3\Characters_F\Common\Suitpacks\suitpack_civilian_F.p3d";
+        class ItemInfo : UniformItem {
+			uniformModel = "-";
+            uniformClass = "tg_opf_gorka_klmk_alt"; // name of linked character
+            containerClass = "Supply50";
+            mass = 50;
+        };
+    };
+
+	class tg_u_gorka_klmk_reverse: Uniform_Base {
+        scope = 2;
+        displayName = "Gorka (KLMK Reverse)";
+		picture = "\tg\opf\Data\logo_small.paa";
+        model = "\A3\Characters_F\Common\Suitpacks\suitpack_civilian_F.p3d";
+        class ItemInfo : UniformItem {
+			uniformModel = "-";
+            uniformClass = "tg_opf_gorka_klmk_reverse"; // name of linked character
+            containerClass = "Supply50";
+            mass = 50;
+        };
+    };
+
+	class tg_u_gorka_klmk_reverse_alt: Uniform_Base {
+        scope = 2;
+        displayName = "Gorka (KLMK Reverse) (alt)";
+		picture = "\tg\opf\Data\logo_small.paa";
+        model = "\A3\Characters_F\Common\Suitpacks\suitpack_civilian_F.p3d";
+        class ItemInfo : UniformItem {
+			uniformModel = "-";
+            uniformClass = "tg_opf_gorka_klmk_reverse_alt"; // name of linked character
+            containerClass = "Supply50";
+            mass = 50;
+        };
+    };
+
+	class tg_u_gorka_partizan: Uniform_Base {
+        scope = 2;
+        displayName = "Gorka (Partizan)";
+		picture = "\tg\opf\Data\logo_small.paa";
+        model = "\A3\Characters_F\Common\Suitpacks\suitpack_civilian_F.p3d";
+        class ItemInfo : UniformItem {
+			uniformModel = "-";
+            uniformClass = "tg_opf_gorka_partizan"; // name of linked character
+            containerClass = "Supply50";
+            mass = 50;
+        };
+    };
+
+	class tg_u_gorka_partizan_alt: Uniform_Base {
+        scope = 2;
+        displayName = "Gorka (Partizan) (alt)";
+		picture = "\tg\opf\Data\logo_small.paa";
+        model = "\A3\Characters_F\Common\Suitpacks\suitpack_civilian_F.p3d";
+        class ItemInfo : UniformItem {
+			uniformModel = "-";
+            uniformClass = "tg_opf_gorka_partizan_alt"; // name of linked character
+            containerClass = "Supply50";
+            mass = 50;
+        };
+    };
+
+	class tg_u_gorka_partizan_autumn: Uniform_Base {
+        scope = 2;
+        displayName = "Gorka (Partizan Autumn)";
+		picture = "\tg\opf\Data\logo_small.paa";
+        model = "\A3\Characters_F\Common\Suitpacks\suitpack_civilian_F.p3d";
+        class ItemInfo : UniformItem {
+			uniformModel = "-";
+            uniformClass = "tg_opf_gorka_partizan_autumn"; // name of linked character
+            containerClass = "Supply50";
+            mass = 50;
+        };
+    };
+
+	class tg_u_gorka_partizan_autumn_alt: Uniform_Base {
+        scope = 2;
+        displayName = "Gorka (Partizan Autumn) (alt)";
+		picture = "\tg\opf\Data\logo_small.paa";
+        model = "\A3\Characters_F\Common\Suitpacks\suitpack_civilian_F.p3d";
+        class ItemInfo : UniformItem {
+			uniformModel = "-";
+            uniformClass = "tg_opf_gorka_partizan_autumn_alt"; // name of linked character
+            containerClass = "Supply50";
+            mass = 50;
+        };
+    };
+
+	/*
+	class tg_u_csat_black: Uniform_Base {
+        scope = 2;
+        displayName = "CSAT Fatigues (Black)";
+		picture = "\tg\opf\Data\logo_small.paa";
+        model = "\A3\Characters_F\Common\Suitpacks\suitpack_civilian_F.p3d";
+        class ItemInfo : UniformItem {
+			uniformModel = "-";
+            uniformClass = "tg_opf_csat_black"; // name of linked character
+            containerClass = "Supply50";
+            mass = 50;
+        };
+    };
+
+	class tg_u_csat_blue: Uniform_Base {
+        scope = 2;
+        displayName = "CSAT Fatigues (Blue)";
+		picture = "\tg\opf\Data\logo_small.paa";
+        model = "\A3\Characters_F\Common\Suitpacks\suitpack_civilian_F.p3d";
+        class ItemInfo : UniformItem {
+			uniformModel = "-";
+            uniformClass = "tg_opf_csat_blue"; // name of linked character
+            containerClass = "Supply50";
+            mass = 50;
+        };
+    };
+
+	class tg_u_csat_yellow: Uniform_Base {
+        scope = 2;
+        displayName = "CSAT Fatigues (CBRN)";
+		picture = "\tg\opf\Data\logo_small.paa";
+        model = "\A3\Characters_F\Common\Suitpacks\suitpack_civilian_F.p3d";
+        class ItemInfo : UniformItem {
+			uniformModel = "-";
+            uniformClass = "tg_opf_csat_yellow"; // name of linked character
+            containerClass = "Supply50";
+            mass = 50;
+        };
+    };
+*/
+};
